split out new node picker
